packages/persistence/mongo/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cqrs-ddd-persistence-mongo"
version = "0.1.0"
description = "MongoDB read-side persistence adapter for CQRS/DDD — projections, Motor async driver"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Myron Gourlis", email = "myrgourlis@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "cqrs-ddd-advanced-core",
    "cqrs-ddd-core",
    "cqrs-ddd-specifications",
    "cqrs-ddd-projections",
    "motor>=3.3.0",
]

[project.optional-dependencies]
advanced = ["cqrs-ddd-advanced-core"]
dispatcher = ["cqrs-ddd-advanced-core"]
dev = ["mongomock-motor>=0.0.0", "pytest-cov>=4.0", "testcontainers>=4.0"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["cqrs_ddd_persistence_mongo*"]
